Conflicts:
	modules/pico_icmp4.h
	modules/pico_ipv4.h

+++ stack/pico_socket.c
@@ -1075,7 +1075,7 @@ static void pico_socket_xmit_next_fragment_setup(struct pico_frame *f, int hdr_o
     f->payload_len = (uint16_t)(f->payload_len + hdr_offset);
     /* set offset in octets */
 #ifdef PICO_SUPPORT_IPFRAG
-    f->frag = short_be((uint16_t)((total_payload_written + (uint16_t)hdr_offset) >> 3u));
+    f->frag = short_be((uint16_t)((uint16_t)(total_payload_written + (uint16_t)hdr_offset) >> 3u));
 #endif
     if (total_payload_written + f->payload_len < len) {
         frag_dbg("FRAG: intermediate fragmented frame %p | len = %u offset = %u\n", f, f->payload_len, short_be(f->frag));
@@ -1497,7 +1497,7 @@ int pico_socket_listen(struct pico_socket *s, int backlog)
     if (PROTO(s) == PICO_PROTO_TCP)
         pico_socket_alter_state(s, PICO_SOCKET_STATE_TCP_SYN_SENT, 0, PICO_SOCKET_STATE_TCP_LISTEN);
 
-    s->max_backlog = backlog;
+    s->max_backlog = (uint16_t)backlog;
 
     return 0;
 }
